Remove Vcc detection and associated variables
Vcc display has been fixed and broken a few times, and
is currently broken.  While we could fix it again  by adding
ADC_MODE(ADC_VCC) as required and adding it to the html
again, there doesn't seem to be any real benefit to showing
it, so just strip it out once and for all.

Signed-off-by: Eric Sandeen <sandeen@sandeen.net>