scale values presented to user & add units to display
Many of the values on the Status tab are just raw from
the evse board, and aren't very user friendly - wattseconds,
for example, which is displayed under a "kWh" heading.

Add a helper function to scale values to a specified precision,
so that instead of presenting, say, 28839184 watt-seconds
as session kWh, we show 8.01kWh.

Signed-off-by: Eric Sandeen <sandeen@sandeen.net>