Hide level min/max for inactive level
If level == 1, we have no values for min/max L2 - and vice
versa.  Hide the irrelevant/unpopulated values.

(level == auto is a different problem; I don't think the
openevse firmware fills in both L1 and L2 in this case,
the RAPI query only returns a L1's values for auto...)

Signed-off-by: Eric Sandeen <sandeen@sandeen.net>