Remove underlines from user-visible states
"estate" is only used as a user-visible EVSE state, spaces
are fine - no_need_for_underlines.  Change the 12 "if"
statements to a case statement while we're at it.

We now need to remove the s.replace(" ", "") though -
That was there to handle mis-parsed tokens with space, and
that has been fixed in previous commits.

Finally, remove the stringification of "estate" in
web_server.cpp; it's already a string.

Signed-off-by: Eric Sandeen <sandeen@sandeen.net>