Set authentication realm to hostname
When the auth dialog box pops up for the user, it
displays the authorization realm along with the
username and password fields, at least on firefox.

The default is "asyncesp" which may be confusing
to some, because that's the name of the library
in use, not the name of the host.

Set it to the device's hostname instead.