'Fix' for auth crashing the ESP
I use the term fix very loosely. What this actually does is allow the StaticFileWebHandler to have the authentication separately configured as with the other handlers and then not configure authentication.

The effect of this is that only the dynamic content asks for authentication. This should be ok but the underlying bug in the new StaticFileWebHandler is still there.