Re-arrange compiling docs, add links to docs
Create a distinct 'Development Guide' to accompany the existing 'User
Guide' and move the sections on compiling/uploading/troubleshooting to
be in this new section.
Expand on some of the notes regarding compiling and uploading.
Add some additional links to relevant documentation throughout the
Development Guide.
Change some instances and references from ESP8266 to be ESP32 relevant.