Remove the legacy API and ENABLE_LEGACY_API flag
This flag was first introduced in 2017, and now that we've moved from
ESP8266 to ESP32, this feels like a good time to get rid of this code.
It is not enabled by default in any build configuration, and a lot of it
was gutted in an earlier commit 48cc8e064.

This endpoint has been merged with the status endpoint since 2017 in
commit 54987e6bdaa.