Only update to the EVSE time if later than the ESP32 time.
On making a connection to the EVSE we read the time as the EVSE module may have a battery backed RTC or even just it has remained on during reboot.

However it also may not have a useful time set, eg when recovering from a power cut and no battery backed RTC. In ths case there is a race condition between getting the time from NTP and the time from the EVSE. So now we only set to the EVSE time if it is later than the local ESP time.

This fixes NTP not automatically working  #194.