Fixed issue with claims that don't set the state
If there was only a single claim that didn't set the state the EVSE manager would not defalut to active, this was only done when there where no claims. This was not a problem for previous modes that always set the state one way or another but for the current shaper it only sets the current and does not set the state, so if it is the only mode in operation and it switched off the charge it would not recover untill something else activated the charge.