Initial rework of divert code
Changed divert to be a MicroTask so we can listen for EVSE state changes, needed to (re)set the timer for the minimum delay timer.

Removed old code to play well with other systems, no longer needed  with the EVSE Manager. This was stopping the divert from stopping under certain conditions.

Various other bits of tidyup.