Be more agressive about clearing the boot lock
There was a check to see if the EVSE was not in the starting state, but it does not exit starting until the boot lock in cleared. Just removed that check as it is not needed as there is a flag to indicate if the boot lock is set.

Fixes #798 “Waiting for initialization” after power loss